backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0
yfinance==0.2.32
backtrader==1.9.78.123
aiohttp>=3.9.0
pandas>=2.0.0
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
